pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "book-capture"
version = "0.1.0"
description = "Traitement de photos de tranches de livres : reconnaissance des titres et auteurs"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["ocr", "livres", "traitement-image", "cli"]
dependencies = [
    "click>=8.0",
    "Pillow>=9.0",
    "pytesseract>=0.3.10",
    "opencv-python-headless>=4.6",
    "openpyxl>=3.0",
    "numpy>=1.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
book-capture = "book_capture.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/book_capture"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
